Permits to build a DFA which returns the "applied distance".
In contrast to reporting the "total" distance needed to match the
given query. This DFA will only report the number of edits which
have been performed so far.

This can be used to "inverse" prefix searches (check if the terms
scanned by the DFA are a prefix of the query).