Permits to determine the number of consumed characters.
Using DFA::offset() for a DFA which returns the "applied distance",
we can now reveal how many characters of the given query have been
consumed. We cannot rely on the number of steps performed, as the edit
distance might have "skipped" or "inserted" a character.

This can be used to "inverse" prefix searches (check if the terms
scanned by the DFA are a prefix of the query).